backend/requirements.txt
python-dotenv

alembic
sqlalchemy
aiosqlite

fastapi[all]
python-multipart
uvicorn

pydantic
pydantic_settings

passlib
python-jose
bcrypt==4.0.0